% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_screen_record.R
\name{end_screen_record}
\alias{end_screen_record}
\alias{is_screen_recording}
\title{End Screen Record}
\usage{
end_screen_record(pid)

is_screen_recording(pid)
}
\arguments{
\item{pid}{Process ID from the screen recording or a
\code{screen_recording}}
}
\value{
A logical vector same length as \code{pid}
}
\description{
End Screen Record
}
