% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_recording_device.R
\name{guess_recording_device}
\alias{guess_recording_device}
\title{Guess the Recording Device for \code{ffmpeg}}
\usage{
guess_recording_device()
}
\value{
A character string of a recording device
}
\description{
Guess the Recording Device for \code{ffmpeg}
}
\examples{
guess_recording_device()
}
