% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_devices.R
\name{list_output_devices}
\alias{list_output_devices}
\title{\code{ffmpeg} Output Devices}
\usage{
list_output_devices(args = c("-devices"))
}
\arguments{
\item{args}{arguments to pass to \link[sys:exec_internal]{sys::exec_internal} to
\code{ffmpeg}}
}
\value{
A \code{data.frame} of device names and descriptions
}
\description{
\code{ffmpeg} Output Devices
}
\examples{
print(list_output_devices())

}
