% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_screen_record.R
\name{print.screen_recording}
\alias{print.screen_recording}
\title{Print method for screen_recording}
\usage{
\method{print}{screen_recording}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for screen_recording
}
\examples{
x = list(process =  processx::process$new("ls"))
class(x) = "screen_recording"
print(x)
}
