% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_script_selection.R
\name{record_script_walk}
\alias{record_script_walk}
\title{Record Script Selection}
\usage{
record_script_walk(output_ext = getOption("record.ext", "gif"))
}
\arguments{
\item{output_ext}{Extension of the output filename}
}
\value{
A filename of the output video
}
\description{
Record Script Selection
}
